IOFile is platform-agnostic (again) Platform-specific binds moved to separate namespace
